- Improved the Equipment creation menu   - Made it so you can see the weapon/sub/apparel's materials as you're building it   - Changed the background color to look sleeker   - Made it so you can see a weapon/sub/apparel's materials in the Collection menu   - Added support for being able to rename objects you've already created (redoing descriptions will be added as well)
- Made it so when you reset the game, it takes you back to the title screen
- Refactored the inventory code to be less error-prone and more intuitive. This also allows for having separate inventories, such as one for testing, one for events, one for demos, one for the final product, etc.
- Added in the nullable code that I deleted for some reason

+++ quantum_unity/Assets/Photon/Quantum/User/InfoAssetAsset.cs
@@ -8,7 +8,7 @@ public abstract partial class InfoAssetAsset : AssetBase
 
     [Space(10)]
 
-    public int Price;
+    public ulong Price;
 
     public Sprite Icon;
     public Sprite Background;
