+++ quantum_unity/Assets/Scripts/Fusion Fighters/Match/MatchController.cs
@@ -92,7 +92,7 @@ public class MatchController : Controller<MatchController>
         if (QuantumRunner.Default.Game.Frames.Verified.TryGetSingleton(out Timer timer))
         {
             int diference = timer.OriginalTime - timer.Time;
-            InventoryController.Instance.GainCurrency((int)(diference * 0.75f));
+            InventoryController.Instance.GainCurrency((ulong)(diference * 0.75f));
         }
     }
 }
